Move message api into its own folder
