+++ pkg/api/message/publishWorker.go
@@ -1,4 +1,4 @@
-package api
+package message
 
 import (
 	"context"
